DEVX-312: separate form state reducer from global one + wrap form component with mandatory provider.
